Added other example to CH10

+++ nucleo-f401RE/src/ch10/main-ex1.c
@@ -41,6 +41,6 @@ int main() {
     }
 }
 
-void delay(unsigned long count) {
+void delay(uint32_t count) {
     while(count--);
 }
