+++ nucleo-f401RE/src/ch10/main-ex4.c
@@ -31,34 +31,34 @@ uint32_t *vector_table[] __attribute__((section(".isr_vector"))) = {
 
 // Begin address for the initialisation values of the .data section.
 // defined in linker script
-extern unsigned int _sidata;
+extern uint32_t _sidata;
 // Begin address for the .data section; defined in linker script
-extern unsigned int _sdata;
+extern uint32_t _sdata;
 // End address for the .data section; defined in linker script
-extern unsigned int _edata;
+extern uint32_t _edata;
 // Begin address for the .bss section; defined in linker script
-extern unsigned int _sbss;
+extern uint32_t _sbss;
 // End address for the .bss section; defined in linker script
-extern unsigned int _ebss;
+extern uint32_t _ebss;
 
 inline void
 __attribute__((always_inline))
-__initialize_data (unsigned int* from, unsigned int* region_begin, unsigned int* region_end)
+__initialize_data (uint32_t* from, uint32_t* region_begin, uint32_t* region_end)
 {
   // Iterate and copy word by word.
   // It is assumed that the pointers are word aligned.
-  unsigned int *p = region_begin;
+  uint32_t *p = region_begin;
   while (p < region_end)
     *p++ = *from++;
 }
 
 inline void
 __attribute__((always_inline))
-__initialize_bss (unsigned int* region_begin, unsigned int* region_end)
+__initialize_bss (uint32_t* region_begin, uint32_t* region_end)
 {
   // Iterate and copy word by word.
   // It is assumed that the pointers are word aligned.
-  unsigned int *p = region_begin;
+  uint32_t *p = region_begin;
   while (p < region_end)
     *p++ = 0;
 }
@@ -89,6 +89,6 @@ int main() {
    	}
 }
 
-void delay(unsigned long count) {
+void delay(uint32_t count) {
     while(count--);
 }
